reorg challenge refactor
