+++ HabitRPG/TableviewCells/Challenges/ChallengeDescriptionTableViewCell.xib
@@ -19,14 +19,14 @@
                 <rect key="frame" x="0.0" y="0.0" width="320" height="195.5"/>
                 <autoresizingMask key="autoresizingMask"/>
                 <subviews>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" ambiguous="YES" text="Challenge Description" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="7yH-Ll-eGO">
-                        <rect key="frame" x="16" y="8" width="169" height="21"/>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Challenge Description" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="7yH-Ll-eGO">
+                        <rect key="frame" x="16" y="8" width="169" height="24"/>
                         <fontDescription key="fontDescription" type="system" pointSize="17"/>
                         <color key="textColor" red="0.30588235289999999" green="0.2901960784" blue="0.3411764706" alpha="1" colorSpace="calibratedRGB"/>
                         <nil key="highlightedColor"/>
                     </label>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" ambiguous="YES" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="fqD-9q-t6S">
-                        <rect key="frame" x="16" y="37" width="288" height="151"/>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="fqD-9q-t6S">
+                        <rect key="frame" x="16" y="40" width="288" height="16"/>
                         <fontDescription key="fontDescription" type="system" pointSize="13"/>
                         <color key="textColor" red="0.30588235289999999" green="0.2901960784" blue="0.3411764706" alpha="1" colorSpace="calibratedRGB"/>
                         <nil key="highlightedColor"/>
@@ -40,12 +40,14 @@
                     </button>
                 </subviews>
                 <constraints>
-                    <constraint firstAttribute="bottom" secondItem="fqD-9q-t6S" secondAttribute="bottom" constant="7.5" id="6dV-fI-k8o"/>
+                    <constraint firstAttribute="bottom" relation="greaterThanOrEqual" secondItem="fqD-9q-t6S" secondAttribute="bottom" constant="7.5" id="6dV-fI-k8o"/>
                     <constraint firstItem="fqD-9q-t6S" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="16" id="AsB-cS-92h"/>
                     <constraint firstItem="fqD-9q-t6S" firstAttribute="top" secondItem="7yH-Ll-eGO" secondAttribute="bottom" constant="8" id="DPO-t5-bed"/>
                     <constraint firstAttribute="trailing" secondItem="fqD-9q-t6S" secondAttribute="trailing" constant="16" id="IVa-LU-IJS"/>
+                    <constraint firstItem="7yH-Ll-eGO" firstAttribute="top" secondItem="vI1-Ad-fpK" secondAttribute="top" id="QMI-e0-WVI"/>
                     <constraint firstItem="vI1-Ad-fpK" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="8" id="c3P-tJ-Fs8"/>
                     <constraint firstItem="7yH-Ll-eGO" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="8" id="jDF-I2-3Mo"/>
+                    <constraint firstItem="7yH-Ll-eGO" firstAttribute="centerY" secondItem="vI1-Ad-fpK" secondAttribute="centerY" id="mtf-yx-orH"/>
                     <constraint firstAttribute="trailing" secondItem="vI1-Ad-fpK" secondAttribute="trailing" constant="16" id="rkm-1D-Q7x"/>
                     <constraint firstItem="7yH-Ll-eGO" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="16" id="s4f-dD-EkZ"/>
                 </constraints>
